cv-service/requirements.txt
fastapi
uvicorn[standard]
torch==2.3.1
torchvision==0.18.1
Pillow
numpy
supervision
groundingdino-py
